research_assistant/pyproject.toml
[project]
name = "research-assistant"
version = "0.1.0"
description = "Multi-agent research assistant powered by Claude and LangGraph"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "typer[all]>=0.12.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "yfinance>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
export = [
    "python-docx",
    "python-pptx",
]

[project.scripts]
research-assistant = "research_assistant.cli:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
